maplib: Move key types to map_types.h
